# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Anime Multiverse Arena API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface DashboardStats {
  totalPlayers: number;
  activeBattles: number;
  totalGuilds: number;
  totalCharacters: number;
  totalBattlesToday: number;
  newPlayersToday: number;
  marketListingsActive: number;
  activeTournaments: number;
  totalGoldInCirculation: number;
  bannedPlayers: number;
}

export interface ActivityItem {
  id: string;
  type: string;
  description: string;
  timestamp: string;
  playerUsername?: string;
}

export interface ChartPoint {
  date: string;
  value: number;
}

export interface BattleTypeStat {
  type: string;
  count: number;
}

export interface RarityCount {
  rarity: string;
  count: number;
}

export interface DashboardCharts {
  battlesPerDay: ChartPoint[];
  newPlayersPerDay: ChartPoint[];
  battleTypeDistribution: BattleTypeStat[];
  topRarityDistribution: RarityCount[];
}

export interface PlayerSummary {
  id: number;
  discordId: string;
  username: string;
  level: number;
  gold: number;
  gems: number;
  pvpRating: number;
  wins: number;
  losses: number;
  isBanned: boolean;
  currentWorld: string;
  createdAt: string;
}

export interface PlayersPage {
  data: PlayerSummary[];
  total: number;
  page: number;
  totalPages: number;
}

export interface PlayerDetail {
  player: PlayerSummary;
  characterCount: number;
  totalDamageDealt: number;
  maxAbyssFloor: number;
  stamina: number;
  maxStamina: number;
  currentTitle?: string;
}

export interface LeaderboardEntry {
  rank: number;
  discordId: string;
  username: string;
  value: number;
  level: number;
}

export interface BanPlayerRequest {
  ban: boolean;
  reason?: string;
}

export interface ModifyPlayerRequest {
  gold?: number;
  gems?: number;
  xp?: number;
  reason?: string;
}

export interface CharacterSummary {
  id: number;
  name: string;
  animeSource: string;
  rarity: string;
  element1: string;
  element2?: string;
  baseHp: number;
  baseAtk: number;
  baseDef: number;
  baseSpd: number;
  isEnabled: boolean;
}

export interface CharactersPage {
  data: CharacterSummary[];
  total: number;
  page: number;
  totalPages: number;
}

export interface GuildSummary {
  id: number;
  name: string;
  tag: string;
  emblem: string;
  level: number;
  memberCount: number;
  maxMembers: number;
  treasury: number;
  totalWins: number;
  totalBossKills: number;
  isOpen: boolean;
  createdAt: string;
}

export interface GuildsPage {
  data: GuildSummary[];
  total: number;
  page: number;
  totalPages: number;
}

export interface GuildMemberSummary {
  discordId: string;
  username: string;
  role: string;
  contribution: number;
  joinedAt: string;
}

export interface GuildDetail {
  guild: GuildSummary;
  members: GuildMemberSummary[];
}

export interface BattleSummary {
  id: number;
  type: string;
  status: string;
  player1DiscordId: string;
  player2DiscordId?: string;
  winnerId?: string;
  turnNumber: number;
  xpEarned?: number;
  goldEarned?: number;
  createdAt: string;
}

export interface BattlesPage {
  data: BattleSummary[];
  total: number;
  page: number;
  totalPages: number;
}

export interface MarketListingSummary {
  id: number;
  sellerUsername: string;
  itemName: string;
  itemRarity: string;
  itemType: string;
  listingType: string;
  price: number;
  quantity: number;
  status: string;
  createdAt: string;
}

export interface MarketPage {
  data: MarketListingSummary[];
  total: number;
  page: number;
  totalPages: number;
}

export type MarketStatsTopSellersItem = {
  username: string;
  salesCount: number;
};

export interface MarketStats {
  totalVolume: number;
  activeListings: number;
  soldToday: number;
  avgPrice: number;
  topSellers: MarketStatsTopSellersItem[];
}

export interface TournamentSummary {
  id: number;
  name: string;
  status: string;
  size: number;
  currentRound: number;
  totalRounds: number;
  prizePool: number;
  winnerUsername?: string;
  createdAt: string;
}

export interface TournamentsPage {
  data: TournamentSummary[];
  total: number;
  page: number;
  totalPages: number;
}

export interface BossSummary {
  id: number;
  name: string;
  title: string;
  animeSource: string;
  element1: string;
  tier: string;
  hp: number;
  isWorldBoss: boolean;
  isAbyssBoss: boolean;
  isEnabled: boolean;
}

export interface BossesPage {
  data: BossSummary[];
  total: number;
  page: number;
  totalPages: number;
}

export interface AdminLogEntry {
  id: number;
  adminUsername: string;
  action: string;
  targetUsername?: string;
  details?: string;
  reason?: string;
  createdAt: string;
}

export interface AdminLogsPage {
  data: AdminLogEntry[];
  total: number;
  page: number;
  totalPages: number;
}

export type BroadcastRequestType = typeof BroadcastRequestType[keyof typeof BroadcastRequestType];


export const BroadcastRequestType = {
  info: 'info',
  warning: 'warning',
  event: 'event',
  maintenance: 'maintenance',
} as const;

export interface BroadcastRequest {
  message: string;
  type: BroadcastRequestType;
}

export type GetDashboardActivityParams = {
limit?: number;
};

export type GetPlayersParams = {
page?: number;
limit?: number;
search?: string;
sort?: GetPlayersSort;
};

export type GetPlayersSort = typeof GetPlayersSort[keyof typeof GetPlayersSort];


export const GetPlayersSort = {
  level: 'level',
  gold: 'gold',
  pvpRating: 'pvpRating',
  wins: 'wins',
  createdAt: 'createdAt',
} as const;

export type GetLeaderboardParams = {
type?: GetLeaderboardType;
limit?: number;
};

export type GetLeaderboardType = typeof GetLeaderboardType[keyof typeof GetLeaderboardType];


export const GetLeaderboardType = {
  level: 'level',
  pvpRating: 'pvpRating',
  wins: 'wins',
  totalDamageDealt: 'totalDamageDealt',
} as const;

export type GetCharactersParams = {
page?: number;
limit?: number;
rarity?: string;
element?: string;
};

export type GetGuildsParams = {
page?: number;
limit?: number;
search?: string;
};

export type GetBattlesParams = {
page?: number;
limit?: number;
type?: string;
status?: string;
};

export type GetMarketListingsParams = {
page?: number;
limit?: number;
status?: string;
};

export type GetTournamentsParams = {
page?: number;
limit?: number;
status?: string;
};

export type GetBossesParams = {
page?: number;
limit?: number;
};

export type GetAdminLogsParams = {
page?: number;
limit?: number;
};