# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Anime Multiverse Arena API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
  /guilds:
    get:
      operationId: getGuilds
      tags: [guilds]
      summary: List all guilds
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Guilds list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildsPage"
  /guilds/{id}:
    get:
      operationId: getGuild
      tags: [guilds]
      summary: Get guild details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Guild detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildDetail"
...
  /admin/logs:
    get:
      operationId: getAdminLogs
      tags: [admin]
      summary: Get admin action logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Admin logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLogsPage"
  /admin/broadcast:
    post:
      operationId: broadcastAnnouncement
      tags: [admin]
      summary: Broadcast a game announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastRequest"
      responses:
        "200":
          description: Broadcasted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        description:
          type: string
        timestamp:
          type: string
        playerUsername:
          type: string
      required: [id, type, description, timestamp]
...
    GuildsPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/GuildSummary"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, totalPages]
    GuildSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        tag:
          type: string
        emblem:
          type: string
        level:
          type: integer
        memberCount:
          type: integer
        maxMembers:
          type: integer
        treasury:
          type: integer
        totalWins:
          type: integer
        totalBossKills:
          type: integer
        isOpen:
          type: boolean
        createdAt:
          type: string
      required: [id, name, tag, emblem, level, memberCount, maxMembers, treasury, totalWins, totalBossKills, isOpen, createdAt]
    GuildDetail:
      type: object
      properties:
        guild:
          $ref: "#/components/schemas/GuildSummary"
        members:
          type: array
          items:
            $ref: "#/components/schemas/GuildMemberSummary"
      required: [guild, members]
    GuildMemberSummary:
      type: object
      properties:
        discordId:
          type: string
        username:
          type: string
        role:
          type: string
        contribution:
          type: integer
        joinedAt:
          type: string
      required: [discordId, username, role, contribution, joinedAt]
...
    AdminLogsPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminLogEntry"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, totalPages]
    AdminLogEntry:
      type: object
      properties:
        id:
          type: integer
        adminUsername:
          type: string
        action:
          type: string
        targetUsername:
          type: string
        details:
          type: string
        reason:
          type: string
        createdAt:
          type: string
      required: [id, adminUsername, action, createdAt]
    BroadcastRequest:
      type: object
      properties:
        message:
          type: string
        type:
          type: string
          enum: [info, warning, event, maintenance]
      required: [message, type]