# lib/api-zod/src/generated/types/guildMemberSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Anime Multiverse Arena API
 * OpenAPI spec version: 0.1.0
 */

export interface GuildMemberSummary {
  discordId: string;
  username: string;
  role: string;
  contribution: number;
  joinedAt: string;
}